add, delete, patch에 대해 login 로직 구현
* localStorage에 있는 authorization을 확인하고 먼저 판결

* 그 다음 req.body에 userId를 담아서 요청을 보낸 뒤 결과 처리

+++ frontend/src/modules/todo/deleteTodo.js
@@ -9,6 +9,8 @@ const confirmSentence = '선택하신 카드를 삭제하겠습니까?';
 
 export default async function deleteTodo(e) {
   if (e.target.dataset.method !== 'delete') return;
+  if (localStorage.getItem('authorization') !== 'true')
+    return alert('쓰기모드가 아닙니다');
   if (!checkConfirm()) return;
 
   const deleteItem = e.target.closest('li');
@@ -19,9 +21,14 @@ export default async function deleteTodo(e) {
   const id = deleteItem.id;
   const ul = e.target.closest('ul');
 
+  const userId = localStorage.getItem('userId');
+
   try {
-    const result = await deleteFetchManager(`${todoApi}/${id}`);
-    if (result.status !== 200) throw new Error();
+    const result = await deleteFetchManager(`${todoApi}/${id}`, { userId });
+    if (result.status !== 200) {
+      if (result.status === 401) throw new Error('쓰기 모드가 아닙니다');
+      else throw new Error('다시 해주세요');
+    }
     deleteItem.previousElementSibling.remove();
     deleteItem.remove();
 
@@ -29,7 +36,7 @@ export default async function deleteTodo(e) {
     const log = makeRemoveLog(title, groupTitle);
     addTodoLog(log);
   } catch (e) {
-    alert('다시 시도해주세요');
+    return alert(e);
   }
 }
